+++ monoruby/src/value.rs
@@ -2054,6 +2054,7 @@ impl Value {
 
     /// Read-only access to the slot array of a `Struct` subclass
     /// instance. Asserts the receiver is a STRUCT-typed RValue.
+    #[allow(dead_code)]
     pub(crate) fn as_struct(&self) -> &StructInner {
         assert_eq!(ObjTy::STRUCT, self.rvalue().ty());
         // SAFETY: The assert ensures this RValue is a Struct instance.
